+++ apps/meteor/client/sidebar/header/index.tsx
@@ -1,61 +1,21 @@
-import { Sidebar } from '@rocket.chat/fuselage';
-import { useUser, useTranslation } from '@rocket.chat/ui-contexts';
+import { FeaturePreview, FeaturePreviewOn, FeaturePreviewOff } from '@rocket.chat/ui-client';
 import type { ReactElement } from 'react';
-import React, { memo } from 'react';
+import React, { lazy, memo } from 'react';
 
-import { useFeaturePreview } from '../../hooks/useFeaturePreview';
-import UserAvatarWithStatus from './UserAvatarWithStatus';
-import UserMenu from './UserMenu';
-import Administration from './actions/Administration';
-import CreateRoom from './actions/CreateRoom';
-import Directory from './actions/Directory';
-import Home from './actions/Home';
-import Login from './actions/Login';
-import Search from './actions/Search';
-import Sort from './actions/Sort';
+const Header = lazy(() => import('./Header'));
+const HeaderPreview = lazy(() => import('./HeaderPreview'));
 
-const HeaderWithData = (): ReactElement => {
-	const t = useTranslation();
-	const user = useUser();
-	const navigationBarEnabled = useFeaturePreview('navigationBar');
-
-	if (navigationBarEnabled) {
-		return (
-			<Sidebar.TopBar.Section>
-				<Sidebar.TopBar.Actions justifyContent='end' width='100%'>
-					<Search title={t('Search')} />
-					{user && (
-						<>
-							<Directory title={t('Directory')} />
-							<Sort title={t('Display')} />
-							<CreateRoom title={t('Create_new')} data-qa='sidebar-create' />
-							{!navigationBarEnabled && <Administration title={t('Administration')} />}
-						</>
-					)}
-					{!user && <Login title={t('Login')} />}
-				</Sidebar.TopBar.Actions>
-			</Sidebar.TopBar.Section>
-		);
-	}
-
+const HeaderWrapper = (): ReactElement => {
 	return (
-		<Sidebar.TopBar.Section>
-			{user ? <UserMenu user={user} /> : <UserAvatarWithStatus />}
-			<Sidebar.TopBar.Actions>
-				<Home title={t('Home')} />
-				<Search title={t('Search')} />
-				{user && (
-					<>
-						<Directory title={t('Directory')} />
-						<Sort title={t('Display')} />
-						<CreateRoom title={t('Create_new')} data-qa='sidebar-create' />
-						<Administration title={t('Administration')} />
-					</>
-				)}
-				{!user && <Login title={t('Login')} />}
-			</Sidebar.TopBar.Actions>
-		</Sidebar.TopBar.Section>
+		<FeaturePreview feature='navigationBar'>
+			<FeaturePreviewOff>
+				<Header />
+			</FeaturePreviewOff>
+			<FeaturePreviewOn>
+				<HeaderPreview />
+			</FeaturePreviewOn>
+		</FeaturePreview>
 	);
 };
 
-export default memo(HeaderWithData);
+export default memo(HeaderWrapper);
