+++ apps/meteor/client/views/root/MainLayout/LayoutWithSidebar.tsx
@@ -1,11 +1,11 @@
 import { Box } from '@rocket.chat/fuselage';
+import { FeaturePreview, FeaturePreviewOff, FeaturePreviewOn } from '@rocket.chat/ui-client';
 import { useLayout, useSetting, useCurrentModal, useRoute, useCurrentRoutePath } from '@rocket.chat/ui-contexts';
 import { PaletteStyleTag } from '@rocket.chat/ui-theming/src/PaletteStyleTag';
 import { SidebarPaletteStyleTag } from '@rocket.chat/ui-theming/src/SidebarPaletteStyleTag';
 import type { ReactElement, ReactNode } from 'react';
 import React, { useEffect, useRef } from 'react';
 
-import { useFeaturePreview } from '../../../hooks/useFeaturePreview';
 import Navbar from '../../../navbar';
 import Sidebar from '../../../sidebar';
 
@@ -17,7 +17,6 @@ const LayoutWithSidebar = ({ children }: { children: ReactNode }): ReactElement 
 	const channelRoute = useRoute('channel');
 	const removeSidenav = embeddedLayout && !currentRoutePath?.startsWith('/admin');
 	const readReceiptsEnabled = useSetting('Message_Read_Receipt_Store_Users');
-	const navigationBarEnabled = useFeaturePreview('navigationBar');
 
 	const firstChannelAfterLogin = useSetting('First_Channel_After_Login');
 
@@ -53,7 +52,14 @@ const LayoutWithSidebar = ({ children }: { children: ReactNode }): ReactElement 
 			<SidebarPaletteStyleTag />
 			{!removeSidenav ? (
 				<>
-					{navigationBarEnabled && <Navbar />}
+					<FeaturePreview feature='navigationBar'>
+						<FeaturePreviewOn>
+							<Navbar />
+						</FeaturePreviewOn>
+						<FeaturePreviewOff>
+							<></>
+						</FeaturePreviewOff>
+					</FeaturePreview>
 					<Sidebar />
 				</>
 			) : null}
