+++ packages/ui-client/src/components/index.ts
@@ -8,3 +8,4 @@ export * from './TooltipComponent';
 export * as UserStatus from './UserStatus';
 export { default as Card } from './Card';
 export * from './Header';
+export * from './FeaturePreview/FeaturePreview';
